feat(impl): implement feature flag spec
Current TestEnvironment is flakey. Noted as the listing spec counts various amounts of flags existing. Always higher than the test registers. Running in isolation fixes it.

This is likely caused by improper disposal of the environment when re-initializing it. TestContainers might also be quirky here.

Leaving this for later as it's outside the scope of this commit