fix: ensure application starts after object store
Add minio service_healthy condition as application dependency.

This is not strictly speaking necessary with the current setup,
in the same way that the database needs to be present (application
performs migration at startup, necessitating it's prescence).

If bucket's are created in the same manner, this will however pose
a similar problem. So to avoid this a dependency is introduced.

This does encourage  some less than perfect practices as bucket
lifetimes are now controlled by the developer directly. It's an
acceptable tradeoff between developer simplicity and governance
requirements. They can come later.